+++ NEWS.md
@@ -1,3 +1,10 @@
+# easyRasch 0.3.9 (2025-09-22)
+
+- Removed output of outfit MSQ statistics since these are not useful in determining item fit/misfit to the Rasch model (Müller, 2020; Johansson, 2025).
+- Added `RIresidcorrG2()` and the corresponding parametric bootstrap to determine a critical value, `RIgetResidcorrG2()`, for the G2 metric to evaluate local independence (Chen & Thissen, 1997).
+  - Using G^2^ is recommended in addition to Yen's Q^3^ to evaluate local dependence as both have strengths and weaknesses.
+- Namespace fix for explicit use of `kableExtra::footnote()` in tables, to avoid issues with `flextable` and potentially other packages.
+ 
 # easyRasch 0.3.8 (2025-09-18)
 
 - Major speedup for `RIbootRestscore()` for PCM (due to moving to `psychotools::PCModel.fit(hessian = FALSE)` instead of `eRm::PCM()`), and minor speed up for RM (due to not calculating SE).